+++ src/engine/time.h
@@ -7,7 +7,7 @@
 namespace now {
 
 inline long current_ms() {
-    return std::chrono::time_point_cast<std::chrono::milliseconds>(
+    return (long) std::chrono::time_point_cast<std::chrono::milliseconds>(
                std::chrono::high_resolution_clock::now())
         .time_since_epoch()
         .count();
