+++ src/network/server.h
@@ -113,7 +113,7 @@ struct Server {
                                  "waiting for run loop continue", true);
                 end = std::chrono::high_resolution_clock::now();
                 duration =
-                    std::chrono::duration_cast<std::chrono::milliseconds>(end -
+                    (float) std::chrono::duration_cast<std::chrono::milliseconds>(end -
                                                                           start)
                         .count();
                 std::this_thread::sleep_for(1ms);
